requirements.txt
flask
easyocr
opencv-python
pillow
